requirements.txt
# Telegram Bot
aiogram==3.3.0
python-dotenv==1.0.0

# Google Sheets API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# API Clients
requests==2.31.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0

# Scheduling
APScheduler==3.10.4

# Logging
loguru==0.7.2



